Rename page and reformat layouts
